requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
progressbar2==3.53.1
python-utils==2.5.6
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
